moved field selection from controller to service
