+++ src/ImportDefinitionsBundle/Resources/config/services.yml
@@ -12,6 +12,10 @@ services:
     import_definition.service.placeholder:
         class: ImportDefinitionsBundle\Service\Placeholder
 
+    import_definition.service.fieldselection:
+        class: ImportDefinitionsBundle\Service\FieldSelection
+        public: true
+
     import_definition.importer:
         class: ImportDefinitionsBundle\Importer\Importer
         arguments:
@@ -337,3 +341,10 @@ services:
         class: ImportDefinitionsBundle\Loader\PrimaryKeyLoader
         tags:
             - { name: import_definition.loader, type: primary_key }
+#
+#    ImportDefinitionsBundle\Controller\ImportDefinitionController:
+#            public: true
+#            parent: '@coreshop.admin_controller.resource_settings'
+#            calls:
+#              - [setFieldSelection, ['@import_definition.service.fieldselection']]
+#            tags: ['controller.service_arguments']
